pyproject.toml
[build-system]
requires = ["scikit-build-core[pyproject]>=0.9.2"]
build-backend = "scikit_build_core.build"

[project]
name = "stable_diffusion_cpp_python"
dynamic = ["version"]
description = "Python bindings for the stable-diffusion.cpp library"
readme = "README.md"
license = { text = "MIT" }
authors = [{ name = "William Murray" }]
maintainers = [{ name = "William Murray" }]
dependencies = ["typing-extensions>=4.5.0", "pillow>=10.2.0"]
keywords = [
    "diffusion",
    "stable-diffusion",
    "python",
    "cpp-bindings",
    "artificial-intelligence",
    "machine-learning",
]
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]

[project.optional-dependencies]
dev = [
    "black>=24.8.0",
    "isort>=5.13.2",
    "pytest>=7.4.4",
    "ffmpeg-python>=0.2.0",
]

[tool.scikit-build]
wheel.packages = ["stable_diffusion_cpp"]
cmake.verbose = true
cmake.minimum-version = "3.21"
minimum-version = "0.5.1"
sdist.include = [".git", "vendor/stable-diffusion.cpp/.git"]

[tool.scikit-build.metadata.version]
provider = "scikit_build_core.metadata.regex"
input = "stable_diffusion_cpp/__init__.py"

[project.urls]
Homepage = "https://github.com/william-murray1204/stable-diffusion-cpp-python"
Issues = "https://github.com/william-murray1204/stable-diffusion-cpp-python/issues"
Documentation = "https://github.com/william-murray1204/stable-diffusion-cpp-python/blob/main/README.md"

[tool.black]
line-length = 130

[tool.isort]
profile = "black"
known_local_folder = ["stable_diffusion_cpp"]
remove_redundant_aliases = true
length_sort = true

[tool.pytest.ini_options]
testpaths = "tests"
